% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmBuildMetricSubspaces.R
\name{dmBuildMetricSubspaces}
\alias{dmBuildMetricSubspaces}
\title{Build metric subspaces for a level}
\usage{
dmBuildMetricSubspaces(dataModelFileName, level, generativeDataFileName)
}
\arguments{
\item{dataModelFileName}{Name of data model file}

\item{level}{Level}

\item{generativeDataFileName}{Name of generative data file}
}
\value{
None
}
\description{
Read a data model and generative data from files,
analyze the contained neural network in the data model for a level,
determine metric subspaces with density values above a level,
add obtained metric subspaces to the data model
and write it to original file.
}
\examples{
\dontrun{
dmBuildMetricSubspaces("dm.bin", 0.5, "gd.bin")}
}
