% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmBuildSubspaces.R
\name{dmBuildSubspaces}
\alias{dmBuildSubspaces}
\title{Build subspaces for a data model for a level}
\usage{
dmBuildSubspaces(dataModelFileName, level, generativeDataFileName)
}
\arguments{
\item{dataModelFileName}{Name of data model file}

\item{level}{Level}

\item{generativeDataFileName}{Name of generative data file}
}
\value{
None
}
\description{
Read a data model and generative data from files,
analyse the trained neural network in a data model for a level,
determine subspaces with density values above the level,
add obtained subspaces with assigned level to the data model
and write it to original file.
}
\examples{
\dontrun{
dmBuildSubspaces("dm.bin", 0.73, "gd.bin")}
}
