% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmEvaluate.R
\name{dmCalculateDensityValue}
\alias{dmCalculateDensityValue}
\title{Calculate a density value for a data record}
\usage{
dmCalculateDensityValue(dataRecord)
}
\arguments{
\item{dataRecord}{List containing a data record}
}
\value{
Normalized density value
}
\description{
Calculate a density value for a data record by evaluating the contained neural network in a data model.
}
\examples{
\dontrun{
dmRead("dm.bin", "gd.bin")
dmCalculateDensityValue(list(4.4, 2.9, 1.4, 0.2))}
}
