% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmBuildSubspaces.R
\name{dmGetAssignedSubspaces}
\alias{dmGetAssignedSubspaces}
\title{Get subspaces in a data model to which a data record is assigned}
\usage{
dmGetAssignedSubspaces(dataRecord)
}
\arguments{
\item{dataRecord}{List of a data record}
}
\value{
List of pairs of level and enumerated subspace
}
\description{
Determine to which pairs of level and enumerated subspace a data record is assigned.
}
\examples{
\dontrun{
dmRead("dm.bin", "gd.bin")
dmGetAssignedSubspaces(list(4.4, 2.9, 1.4, 0.3))}
}
