% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmBuildMetricSubspaces.R
\name{dmGetContainedInMetricSubspaces}
\alias{dmGetContainedInMetricSubspaces}
\title{Get metric subspaces in which a data record is contained}
\usage{
dmGetContainedInMetricSubspaces(dataRecord)
}
\arguments{
\item{dataRecord}{List of a data record}
}
\value{
List of list containing level and label of metric subspaces
}
\description{
Determine in which metric subspaces in a data model a data record is contained.
}
\examples{
\dontrun{
dmRead("dm.bin", "gd.bin")
dmGetContainedInMetricSubspaces(list(4.4, 2.9, 1.4, 0.2))}
}
