% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmGetMetricSubspaceProperties}
\alias{dmGetMetricSubspaceProperties}
\title{Get metric subspace properties for a level}
\usage{
dmGetMetricSubspaceProperties(level)
}
\arguments{
\item{level}{Level for metric subspaces}
}
\value{
List of list containing label and size of contained generative data for metric subspaces
}
\description{
Get properties of metric subspaces in a data model for a level.
}
\examples{
\dontrun{
dmRead("dm.bin", "gd.bin")
dmGetMetricSubspaceProperties(0.73)}
}
