% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmGetNumberOfSubspaces}
\alias{dmGetNumberOfSubspaces}
\title{Get number of subspaces for a level}
\usage{
dmGetNumberOfSubspaces(level)
}
\arguments{
\item{level}{Level}
}
\value{
Number of subspaces
}
\description{
Get number of subspaces in a data model for a level.
}
\examples{
\dontrun{
dmRead("dm.bin", "gd.bin")
dmGetNumberOfSubspaces(0.73)}
}
