% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmPlotSubspaces.R
\name{dmPlotEvaluateDataSourceParameters}
\alias{dmPlotEvaluateDataSourceParameters}
\title{Specify plot parameters for evaluated data source}
\usage{
dmPlotEvaluateDataSourceParameters(colour = "blue")
}
\arguments{
\item{colour}{Colour for data points of evaluaded data source}
}
\value{
List of plot parameters for evaluated data source
}
\description{
Specify plot parameters for evaluated data source passed to function dmPlotSubspaces().
}
\examples{
\dontrun{
dmPlotEvaluateDataSourceParameters()}
}
