% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmPlotSubspaces.R
\name{dmPlotGenerativeDataParameters}
\alias{dmPlotGenerativeDataParameters}
\title{Specify plot parameters for generative data}
\usage{
dmPlotGenerativeDataParameters(percent = 10, colour = "red")
}
\arguments{
\item{percent}{Percentr of randomly selected data points of generative data}

\item{colour}{Colour for data points of generative data}
}
\value{
List of plot parameters for generative data
}
\description{
Specify plot parameters for generative data passed to function dmPlotSubspaces().
}
\examples{
\dontrun{
dmPlotGenerativeDataParameters(5)}
}
