% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmPlotSubspaces.R
\name{dmPlotSubspaceParameters}
\alias{dmPlotSubspaceParameters}
\title{Specify plot parameters for subspaces in a data model}
\usage{
dmPlotSubspaceParameters(percent = 100, boundary = TRUE, colour = "green")
}
\arguments{
\item{percent}{Percent of randomly selected data points of generative data assigned to subspaces}

\item{boundary}{Boolean value indicating if only data points of subspace boundaries should be selected}

\item{colour}{Colour for data points of generative data assigned to subspaces}
}
\value{
List of plot parameters for subspaces
}
\description{
Specify plot parameters for subspaces in a data model passed to function dmPlotSubspaces().
}
\examples{
\dontrun{
dmPlotSubspaceParameters(50)}
}
