% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmBuildSubspaces.R
\name{dmRemoveSubspaces}
\alias{dmRemoveSubspaces}
\title{Remove subspaces from a data model for a level}
\usage{
dmRemoveSubspaces(dataModelFileName, level)
}
\arguments{
\item{dataModelFileName}{Name of data model file}

\item{level}{Level}
}
\value{
None
}
\description{
Read a data model from file,
remove subspaces from the data model for a level
and write it to original file.
}
\examples{
\dontrun{
dmRead("dm.bin", "gd.bin")
dmRemoveSubspaces("dm.bin", 0.73)}
}
