% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmGetLevels}
\alias{dmGetLevels}
\title{Get levels for metric subspaces}
\usage{
dmGetLevels()
}
\value{
Vector of levels
}
\description{
Get levels for metric subspaces in a data model.
}
\examples{
\dontrun{
dmRead("dm.bin", "gd.bin")
dmGetLevels()}
}
