% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmPlotMetricSubspaces.R
\name{dmPlotEvaluateDataSourceParameters}
\alias{dmPlotEvaluateDataSourceParameters}
\title{Specify plot parameters for evaluated data source}
\usage{
dmPlotEvaluateDataSourceParameters(level = 0, color = "blue")
}
\arguments{
\item{level}{Level for evaluation}

\item{color}{Color for data points of evaluaded data source}
}
\value{
List of plot parameters for evaluated data source
}
\description{
Specify plot parameters for evaluated data source passed to dmPlotMetricSubspaces().
}
\examples{
\dontrun{
dmPlotEvaluateDataSourceParameters()}
}
