% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmEvaluate.R
\name{dmRead}
\alias{dmRead}
\title{Read a data model and generative data}
\usage{
dmRead(dataModelFileName, generativeDataFileName)
}
\arguments{
\item{dataModelFileName}{Name of data model file}

\item{generativeDataFileName}{Name of generative data file}
}
\value{
None
}
\description{
Read a data model and generative data from files.
This function has to be called before calling API functions when
file names for a data model and gernerative data are not passed to functions directly.
}
\examples{
\dontrun{
dmRead("dm.bin", "gd.bin")}
}
