% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmTrain.R
\name{dmTrain}
\alias{dmTrain}
\title{Train a neural network which approximates density values for a data source}
\usage{
dmTrain(
  dataModelFileName,
  dataSourceFileName,
  generativeDataFileName,
  numberOfIterations
)
}
\arguments{
\item{dataModelFileName}{Name of data model file}

\item{dataSourceFileName}{Name of data source file}

\item{generativeDataFileName}{Name of generative data file}

\item{numberOfIterations}{Number of iterations.}
}
\value{
None
}
\description{
Read a data source and generative data from files,
train a neural network
which approximates density values for a data source in iterative training steps,
create a data model containing the trained neural network 
and write it to a file in binary format.
}
\examples{
\dontrun{
dmTrain("dm.bin", "ds.bin", "gd.bin", 10000)}
}
